Fixed unfair "Chet starts in Penalty Box" bug in Ruby example
In Ruby, zero is truthy. The penalty box array was initialized with all
zeros.  Because of the off-by-one error when adding players to the game,
this would mean that Chet would start in the penalty box.

The bug was difficult to spot and was only discovered when we tried to
"fix" the penalty box at a legacy coderetreat when we inadvertently
fixed the bug (we couldn't understand why our output changed drastically
when we ran our Golden Master). The bug also only exists in the Ruby
version (due to Ruby treating zero as equal to true).